# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trada AI trading platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent details and performance history
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── RENTALS ──────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List pending withdrawal requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, approved, broadcasted, completed, failed, rejected]
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalList"
  /admin/withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a withdrawal request
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
  /admin/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a withdrawal request
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectWithdrawalBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
  /admin/rentals/{rentalId}/pause:
    post:
      operationId: pauseRental
      tags: [admin]
      summary: Pause a rental (admin)
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /admin/rentals/{rentalId}/resume:
    post:
      operationId: resumeRental
      tags: [admin]
      summary: Resume a paused rental (admin)
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /admin/rentals/{rentalId}/reset-circuit-breaker:
    post:
      operationId: resetCircuitBreaker
      tags: [admin]
      summary: Reset circuit breaker for a rental (admin)
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Circuit breaker reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /admin/alerts:
    get:
      operationId: listAdminAlerts
      tags: [admin]
      summary: List system alerts
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertList"
  /admin/alerts/{alertId}/resolve:
    post:
      operationId: resolveAlert
      tags: [admin]
      summary: Resolve a system alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    PerformancePoint:
      type: object
      properties:
        date:
          type: string
        balance:
          type: number
        pnl:
          type: number
      required: [date, balance, pnl]

    # ── WALLET ──────────────────────────────────────────────────────────────
...
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        currency:
          type: string
        address:
          type: string
        network:
          type: string
        status:
          type: string
          enum: [requested, approved, broadcasted, completed, failed, rejected]
        rejectionReason:
          type: string
          nullable: true
        txHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, amount, currency, address, network, status, createdAt, updatedAt]
    WithdrawalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WithdrawalRequest"
        total:
          type: integer
      required: [items, total]
    RejectWithdrawalBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]

    # ── AGENTS ──────────────────────────────────────────────────────────────
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        symbol:
          type: string
        strategy:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        rentalPricePerDay:
          type: number
        minBalance:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        avgReturn:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - symbol
        - strategy
        - riskLevel
        - rentalPricePerDay
        - minBalance
        - winRate
        - totalTrades
        - avgReturn
        - isActive
        - createdAt
...
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/Agent"
        - type: object
          properties:
            recentPerformance:
              type: array
              items:
                $ref: "#/components/schemas/PerformancePoint"
          required: [recentPerformance]

    # ── RENTALS ─────────────────────────────────────────────────────────────
    Rental:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        status:
          type: string
          enum: [pending, active, paused, expired, cancelled]
        durationDays:
          type: integer
        totalCost:
          type: number
        startAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        circuitBreakerTripped:
          type: boolean
        consecutiveFailures:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - agentId
        - agentName
        - status
        - durationDays
        - totalCost
        - circuitBreakerTripped
        - consecutiveFailures
        - createdAt
...
    AdminUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required: [items, total]
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        walletBalance:
          type: number
        activeRentals:
          type: integer
        totalTrades:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, walletBalance, activeRentals, totalTrades, createdAt]
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [circuit_breaker_tripped, signal_generation_failed, trade_execution_failed, unusual_withdrawal, cron_stalled]
        severity:
          type: string
          enum: [info, warning, critical]
        title:
          type: string
        message:
          type: string
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        resolved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, type, severity, title, message, resolved, createdAt]
    AlertList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
      required: [items, total]