# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trada AI trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Agent } from "./agent";
import type { PerformancePoint } from "./performancePoint";

export type AgentDetail = Agent & {
  recentPerformance: PerformancePoint[];
};