# lib/api-zod/src/generated/types/withdrawalList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trada AI trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WithdrawalRequest } from "./withdrawalRequest";

export interface WithdrawalList {
  items: WithdrawalRequest[];
  total: number;
}